chore: implement visual review architectural improvements and cleanup (#2885)
