+++ .github/workflows/self-healing.yml
@@ -72,7 +72,6 @@ jobs:
           rm -rf dev-tools
           cp -r trusted-tools-temp/dev-tools dev-tools
           rm -rf trusted-tools-temp
-          chmod +x dev-tools/low-compute-review.sh
 
       - name: Setup Node and pnpm
         uses: ./.github/actions/setup-node-pnpm
