+++ dev-tools/dev_tools_sdk/config.py
@@ -13,9 +13,6 @@ class ProjectConfig:
     github_token_env: str = "GITHUB_TOKEN"
     gh_token_env: str = "GH_TOKEN"
     use_gemini_fallback: bool = True
-
-
-
     jules_api_url: str | None = None
 
 
@@ -49,9 +46,6 @@ def load_project_config(path: str | Path = "dev-tools/project_config.json") -> P
         github_token_env=raw.get("github_token_env", "GITHUB_TOKEN"),
         gh_token_env=raw.get("gh_token_env", "GH_TOKEN"),
         use_gemini_fallback=fallback_val if fallback_val is not None else _coerce_bool(raw.get("use_gemini_fallback"), True),
-
-
-
         jules_api_url=raw.get("jules_api_url"),
     )
