+++ dev-tools/project_config.json
@@ -6,7 +6,6 @@
   "monolithic_pr_threshold": 3,
   "base_branch": "origin/main",
   "use_gemini_fallback": false,
-
   "content_scopes": {
     "resources": "content/resources/",
     "posts": "content/posts/",
