+++ src/config/devai-assets.ts
@@ -21,6 +21,18 @@ export const DEVAI_ASSETS: DevAIAsset[] = [
     toolId: TOOL_ID_SDLC_PROFILER
   },
   {
+    path: 'dev-tools/mergellama.py',
+    label: 'PR Auditor',
+    description: 'Autonomous LLM-based code review agent.',
+    toolId: TOOL_ID_GITOPS_PR_REVIEWER
+  },
+  {
+    path: 'dev-tools/ollama_reviewer.py',
+    label: 'Local LLM Reviewer',
+    description: 'Local-first model execution for privacy-safe reviews.',
+    toolId: TOOL_ID_GITOPS_PR_REVIEWER
+  },
+  {
     path: 'dev-tools/td_cli.py',
     label: 'CLI Manager',
     description: 'Unified command-line interface for DevAI operations.',
