+++ src/constants/visual-viewports.ts
@@ -1,0 +1,7 @@
+export const VIEWPORTS = [
+  { name: 'Desktop', width: 1440, height: 900, suffix: '' },
+  { name: 'Laptop', width: 1280, height: 800, suffix: '-laptop' },
+  { name: 'Tablet', width: 768, height: 1024, suffix: '-tablet' },
+  { name: 'Mobile', width: 390, height: 844, suffix: '-mobile' },
+  { name: 'Ultrawide', width: 1920, height: 1080, suffix: '-ultrawide' },
+];
