+++ src/features/ux-auditor/useUXAuditor.ts
@@ -50,7 +50,8 @@ async function withRetry<T>(fn: () => Promise<T>, maxRetries = 3, baseDelay = 10
 }
 
 // --- Configuration & Constants ---
-const apiKey = import.meta.env.VITE_OPENAI_API_KEY || import.meta.env.VITE_GEMINI_API_KEY || sessionStorage.getItem('ux-auditor-api-key') || "";
+const env = (typeof import.meta !== 'undefined' && import.meta.env ? import.meta.env : {}) as Record<string, string>;
+const apiKey = env.VITE_OPENAI_API_KEY || env.VITE_GEMINI_API_KEY || (typeof sessionStorage !== 'undefined' ? sessionStorage.getItem('ux-auditor-api-key') : "") || "";
 declare const __app_id: string | undefined;
 declare const __firebase_config: string | undefined;
 declare const __initial_auth_token: string | undefined;
@@ -58,11 +59,9 @@ declare const __initial_auth_token: string | undefined;
 const appId = typeof __app_id !== 'undefined' ? __app_id : 'ux-auditor-v2';
 const firebaseConfig = typeof __firebase_config !== 'undefined' ? JSON.parse(__firebase_config) : null;
 
-export const VIEWPORTS = [
-  { name: 'Mobile', width: 375, height: 667 },
-  { name: 'Tablet', width: 768, height: 1024 },
-  { name: 'Desktop', width: 1440, height: 900 }
-];
+import { VIEWPORTS } from '@/constants/visual-viewports';
+
+export { VIEWPORTS };
 
 export interface Improvement {
   element: string;
